Move verseListLoadingIndicator related functions to verseListController
